drawmenu: do not left-pad vertical menu items by prompt size
there's no reason for this, seems like a bug to me

we do still leave a little padding equivalent to lrpad / 2 on each side.
just so it sticks out from the input line slightly.  we'll see if we
ever want to eliminate this space later and just go flush with the sides
and bar color all the way to margin.  maybe.